Fix remaining clang compile warnings

+++ cl_video.c
@@ -199,7 +199,7 @@ static void LoadSubtitles( clvideo_t *video, const char *subtitlesfile )
 	if (numsubs > 0) // make true len for prev sub, autofix overlapping subtitles
 	{
 		if (video->subtitle_end[numsubs-1] <= 0)
-			video->subtitle_end[numsubs-1] = 99999999; // fixme: make it end when video ends?
+			video->subtitle_end[numsubs-1] = (float)99999999; // fixme: make it end when video ends?
 		else
 			video->subtitle_end[numsubs-1] = video->subtitle_start[numsubs-1] + video->subtitle_end[numsubs-1];
 	}
